last version

+++ barseek-bar-mservice/src/main/java/com/example/barseek_bar_mservice/controller/DrinkController.java
@@ -2,6 +2,7 @@ package com.example.barseek_bar_mservice.controller;
 
 
 import com.example.barseek_bar_mservice.dto.UpdateDrinkRequest;
+import com.example.barseek_bar_mservice.exception.customExceptions.UnauthorizedAccessException;
 import com.example.barseek_bar_mservice.model.entity.Bar;
 import com.example.barseek_bar_mservice.model.entity.Drink;
 import com.example.barseek_bar_mservice.security.UserPrincipal;
@@ -27,9 +28,13 @@ public class DrinkController {
     @PostMapping("/{barId}/drinks")
     public ResponseEntity<String> addNew(@PathVariable("barId") Long barId,
                                          @RequestBody Drink drink,
-                                         @AuthenticationPrincipal UserPrincipal principal
+                                         @RequestHeader(value = "X-User-Id") String userId,
+                                         @RequestHeader(value = "X-User-Roles") String roles
     ) {
-            Long ownerId = Long.parseLong(principal.getUsername());
+
+            if(!(roles.contains("bar"))) throw new UnauthorizedAccessException("User is not a bar owner!");
+
+            Long ownerId = Long.parseLong(userId);
             Drink newDrink = drinkService.addNewDrink(barId,drink,ownerId);
             return new ResponseEntity<>("New drink created with name : " + newDrink.getName(),HttpStatus.CREATED);
     }
@@ -45,8 +50,13 @@ public class DrinkController {
     @DeleteMapping("/{barId}/drinks/{drinkId}")
     public ResponseEntity<String> deleteById(@PathVariable("barId") Long barId,
                                              @PathVariable("drinkId") Long drinkId,
-                                             @AuthenticationPrincipal UserPrincipal principal) {
-            Long ownerId = Long.parseLong(principal.getUsername());
+                                             @RequestHeader(value = "X-User-Id") String userId,
+                                             @RequestHeader(value = "X-User-Roles") String roles
+    ) {
+
+            if(!(roles.contains("bar"))) throw new UnauthorizedAccessException("User is not a bar owner!");
+
+            Long ownerId = Long.parseLong(userId);
             drinkService.deleteDrinkById(barId,drinkId,ownerId);
             return new ResponseEntity<>("Drink was deleted id :" + drinkId,HttpStatus.OK);
     }
@@ -66,9 +76,13 @@ public class DrinkController {
             @PathVariable("barId") Long barId,
             @PathVariable("drinkId") Long drinkId,
             @RequestBody UpdateDrinkRequest updateDrinkRequest,
-            @AuthenticationPrincipal UserPrincipal principal
+            @RequestHeader(value = "X-User-Id") String userId,
+            @RequestHeader(value = "X-User-Roles") String roles
     ) {
-            Long ownerId = Long.parseLong(principal.getUsername());
+
+            if(!(roles.contains("bar"))) throw new UnauthorizedAccessException("User is not a bar owner!");
+
+            Long ownerId = Long.parseLong(userId);
             Drink newDrink = drinkService.updateDrinkById(barId,drinkId,updateDrinkRequest,ownerId);
             return new ResponseEntity<>("Drink was saved and updated, name : " + newDrink.getName(),HttpStatus.OK);
     }
